refactor: move runners to directories of their level
 (config-info -> config/, dataset-info -> dataset/)
