+++ services/worker/tests/job_runners/config/test_info.py
@@ -12,7 +12,7 @@ from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 
 from worker.config import AppConfig
-from worker.job_runners.config_info import (
+from worker.job_runners.config.info import (
     ConfigInfoJobRunner,
     PreviousStepFormatError,
     PreviousStepStatusError,
