+++ services/worker/tests/job_runners/dataset/test_info.py
@@ -12,15 +12,15 @@ from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 
 from worker.config import AppConfig
-from worker.job_runners.dataset_info import (
+from worker.job_runners.dataset.info import (
     DatasetInfoJobRunner,
     PreviousStepFormatError,
     PreviousStepStatusError,
 )
 from worker.utils import PreviousJob
 
-from .test_config_info import CONFIG_INFO_1, CONFIG_INFO_2, DATASET_INFO_OK
-from .utils import UpstreamResponse
+from ..config.test_info import CONFIG_INFO_1, CONFIG_INFO_2, DATASET_INFO_OK
+from ..utils import UpstreamResponse
 
 
 @pytest.fixture(autouse=True)
